+++ Modules/_testcapi/parts.h
@@ -4,7 +4,23 @@
 // Always enable assertions
 #undef NDEBUG
 
+// The _testcapi extension tests the public C API: header files in Include/ and
+// Include/cpython/ directories. The internal C API must not be tested by
+// _testcapi: use _testinternalcapi for that.
+//
+// _testcapi C files can built with the Py_BUILD_CORE_BUILTIN macro defined if
+// one of the Modules/Setup files asks to build _testcapi as "static"
+// (gh-109723).
+//
+// The Visual Studio projects builds _testcapi with Py_BUILD_CORE_MODULE.
+#undef Py_BUILD_CORE_MODULE
+#undef Py_BUILD_CORE_BUILTIN
+
 #include "Python.h"
+
+#ifdef Py_BUILD_CORE
+#  error "_testcapi must test the public Python C API, not the internal C API"
+#endif
 
 int _PyTestCapi_Init_Vectorcall(PyObject *module);
 int _PyTestCapi_Init_Heaptype(PyObject *module);
