+++ Modules/_testcapi/pyatomic.c
@@ -4,10 +4,6 @@
  * This only tests basic functionality, not any synchronizing ordering.
  */
 
-/* Always enable assertions */
-#undef NDEBUG
-
-#include "Python.h"
 #include "parts.h"
 
 // We define atomic bitwise operations on these types
